Player scene and movement
movement and animations done. attack and shield animations are done. When everything else is ready we can then add to the script that enemy damage is not taken when shield is up and another thing still missing is the sword collision, but other than that everything is set up for the movement.